Refactoring notification of successfully copy resource link to be through output emitter.
