+++ ontotext-yasgui-web-component/src/components/ontotext-yasgui-web-component/readme.md
@@ -43,7 +43,6 @@ yasgui can be tweaked using the values from the configuration.
 | `createSavedQuery`     | Event emitted when saved query payload is collected and the query should be saved by the component client.                                              | `CustomEvent<SaveQueryData>`                                           |
 | `deleteSavedQuery`     | Event emitted when a saved query should be deleted. In result the client must perform a query delete.                                                   | `CustomEvent<SaveQueryData>`                                           |
 | `loadSavedQueries`     | Event emitted when saved queries is expected to be loaded by the component client and provided back in order to be displayed.                           | `CustomEvent<boolean>`                                                 |
-| `notify`               | Event emitted when there is a message which the client might want to show to the user or handle in some other way.                                      | `CustomEvent<NotificationMessage>`                                     |
 | `output`               | Event emitter used to send message to the clients of component.                                                                                         | `CustomEvent<OutputEvent>`                                             |
 | `queryExecuted`        | Event emitted when before query to be executed.                                                                                                         | `CustomEvent<{ query: string; request: Request; queryMode: string; }>` |
 | `queryResponse`        | Event emitted when after query response is returned.                                                                                                    | `CustomEvent<{ duration: number; }>`                                   |
