+++ ontotext-yasgui-web-component/src/models/event.ts
@@ -27,18 +27,3 @@ export class InternalShowResourceCopyLinkDialogEvent {
 
   constructor(public copyLink: string) {}
 }
-
-export enum NotificationMessageType {
-  SUCCESS = 'success',
-  ERROR = 'error',
-  WARNING = 'warning'
-}
-
-export enum NotificationMessageCode {
-  RESOURCE_LINK_COPIED_SUCCESSFULLY = 'resource_link_copied_successfully'
-}
-
-export class NotificationMessage {
-  constructor(public code: NotificationMessageCode, public type: NotificationMessageType, public message: string) {
-  }
-}
